main: add IO500_MINWRITE constant
Add a constant for the 300s minimum runtime, rather than
hard-coding it in several places in the code.  It will
eventually also needed for Student Cluster Competition.

Use it to print "-invalid" markers for any subtests that
didn't run for a long enough time.  That should never
happen, but it makes the bad result(s) easier to see.

Signed-off-by: Andreas Dilger <adilger@dilger.ca>