io500: allow combined app/script runs and results
Add a new "--timestamp" parameter to the binary app so that
it is possible to run both the script and binary app tests
with the same timestamp, but use a different directory suffix.

The order of the io500 script vs. binary app runs is random.

Make the example datadir and resultdir names consistent across
the provided .ini files for simplicity's sake.

Tar both results directories into one package for submission.
This results in a tarball named io500-$HOSTNAME-$timestamp.tgz
so that there is some chance of avoiding confusion.  The tarball
has two subdirectories - $timestamp-scr and $timestamp-app with
the respective script and binary app test run results in it.

Signed-off-by: Andreas Dilger <adilger@dilger.ca>