io500.sh: simplify configuration for users
Move the setup_paths() function to the top of the io500.sh script,
since this should be the only thing that users need to change here.

Add a config-full.ini example file that contains all of the
parameters from "io500 --list" so users don't need to know that.

Fix stringification of IO500_MINWRITE in "io500 --list" output.

Fixes: 725acb363761 ("main: add IO500_MINWRITE constant")
Signed-off-by: Andreas Dilger <adilger@dilger.ca>