io500.sh: update for isc20 submissions
Update revisions to be used for ISC20 submissions to "io500-isc20".
These do not exist yet, but will be created with the required tag
included, so that users can check out the right version.

Include io500.sh and config.ini files in the result directory/tarball.

Move the remaining configuration of the io500.sh script to the top
of the file so that it is easier for users to edit.

Signed-off-by: Andreas Dilger <adilger@dilger.ca>